examples/tdmq_pulsar_with_scf/consumer/requirements.txt
pulsar-client